+++ py_module/module.cpp
@@ -1,0 +1,6 @@
+#include "python/python_module.hpp"
+
+PYBIND11_MODULE(glaxnimate, glaxnimate_module)
+{
+    register_py_module(glaxnimate_module);
+}