+++ src/core/python/python_module.hpp
@@ -1,0 +1,3 @@
+#include <pybind11/pybind11.h>
+
+void register_py_module(pybind11::module& m);